split enum match validation
