+++ src/typechecker/patterns.rs
@@ -7,6 +7,7 @@ use super::TypeChecker;
 
 mod match_validation;
 mod match_validation_bool;
+mod match_validation_enum;
 
 impl TypeChecker {
     /// Look up the payload type for a specific enum variant.
